Reorganiza ViewModels e interfaces para Domínio/Listas
Refatora classes ArquivoItem, CampoErroWpf, ListViewDxf e TabViewModel para namespaces TNKDxf.Dominio.Listas/Entidades, promovendo melhor separação de responsabilidades. Remove interfaces não utilizadas (IDataExtraido, IListaExtraida, IRevisoesExtraidas) e MockTeklaHandler.cs. Atualiza o .csproj e usings conforme a nova estrutura. Nenhuma lógica foi alterada, apenas reorganização estrutural.
