--- TNKDxf/Dominio/Listas/IDataExtraido.cs
@@ -1,11 +1,0 @@
-﻿namespace TNKDxf.Dominio.Coletores
-{
-    public interface IDataExtraido
-    {
-        //bool PodeInserirBlocos();
-
-        //void Processar();
-
-        //void ApagarSelecao();
-    }
-}
