--- TNKDxf/Dominio/Listas/IListaExtraida.cs
@@ -1,8 +1,0 @@
-﻿using TNKDxf.Dominio.Coletores;
-
-namespace TNKDxf.Dominio.Listas
-{
-    public interface IListaExtraida : IDataExtraido
-    {
-    }
-}
