+++ TNKDxf/TNKDxf.csproj
@@ -107,6 +107,7 @@
     <Compile Include="BoolToAbrirFecharConverter.cs" />
     <Compile Include="Dominio\Dxfs\ArquivoDxf.cs" />
     <Compile Include="Dominio\Dxfs\ColecaoDxfs.cs" />
+    <Compile Include="Dominio\Entidades\Class1.cs" />
     <Compile Include="ExportadoraDxf.cs" />
     <Compile Include="Handles\AvaliadorDesenhos.cs" />
     <Compile Include="Handles\CommandResult.cs" />
@@ -114,17 +115,16 @@
     <Compile Include="Handles\HandleCriacaoDxfs.cs" />
     <Compile Include="Handles\Notification.cs" />
     <Compile Include="Infra\IServicoEnvioDesenhos.cs" />
-    <Compile Include="Infra\MockTeklaHandler.cs" />
     <Compile Include="Infra\ServicoEnvioDesenhos.cs" />
     <Compile Include="Infra\TeklaHandler.cs" />
     <Compile Include="MainViewModel.cs" />
     <Compile Include="Ncs\NcParaDxf.cs" />
     <Compile Include="RelayCommand.cs" />
     <Compile Include="ViewModelBase.cs" />
-    <Compile Include="ViewModel\ArquivoItem.cs" />
-    <Compile Include="ViewModel\CampoErroWpf.cs" />
-    <Compile Include="ViewModel\ListViewDxf.cs" />
-    <Compile Include="ViewModel\TabViewModel.cs" />
+    <Compile Include="Dominio\Listas\ArquivoItem.cs" />
+    <Compile Include="Dominio\Entidades\CampoErroWpf.cs" />
+    <Compile Include="Dominio\Listas\ListViewDxf.cs" />
+    <Compile Include="Dominio\Entidades\TabViewModel.cs" />
     <Page Include="Assets\Icons\FlowDrawingIcon.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -136,12 +136,9 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Dominio\Listas\IDataExtraido.cs" />
     <Compile Include="Dominio\Enumeracoes\Enumeration.cs" />
     <Compile Include="Dominio\Enumeracoes\FormatoType.cs" />
     <Compile Include="Dominio\Extensoes\ExtensionsDouble.cs" />
-    <Compile Include="Dominio\Listas\IListaExtraida.cs" />
-    <Compile Include="Dominio\Listas\IRevisoesExtraidas.cs" />
     <Compile Include="Dominio\ObjetosValor\CampoTexto.cs" />
     <Compile Include="Dominio\ObjetosValor\Codigo.cs" />
     <Compile Include="Dominio\ObjetosValor\Descricao.cs" />
@@ -203,9 +200,7 @@
       <Install>false</Install>
     </BootstrapperPackage>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Infra\Dtos\" />
-  </ItemGroup>
+  <ItemGroup />
   <ItemGroup>
     <Resource Include="Fonts\README.txt" />
   </ItemGroup>
