+++ TNKDxf/ViewModelBase.cs
@@ -1,7 +1,5 @@
-﻿using System.Collections.ObjectModel;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.Runtime.CompilerServices;
-using TNKDxf.ViewModel;
 
 namespace TNKDxf
 {
